+++ getMemory/getMemory.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+    "fmt"
+    "github.com/pbnjay/memory"
+)
+
+func main() {
+    fmt.Println(memory.TotalMemory())
+}