+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## 3.2.6.0
 
--   Add featured packs tab
+- Add featured packs tab
+- Add message to accounts tab with info on what account to use