+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.2.6.1
 
 - Warn if users aren't on at least version 101 of Java 8
+- Change authlib version to 1.5.21 to fix issues with needing Java 8