+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
-## 3.2.6.1
+## 3.2.7.0
 
 - Warn if users aren't on at least version 101 of Java 8
 - Change authlib version to 1.5.21 to fix issues with needing Java 8
 - Don't allow empty password when launching Minecraft without saved password
 - Fix view mods button not working
 - Remove old xml code from 2014 no longer user
+- Make launcher compatable with Java 9 and newer